backend/requirements.txt
fastapi==0.92.0
SQLAlchemy==2.0.4
uvicorn==0.20.0
black==23.1.0
flake8==6.0.0
isort==5.12.0
pytest==7.2.1
mysqlclient==2.1.1
Authlib==1.2.0
httpx==0.24.1
itsdangerous==2.1.2
shortuuid==1.0.11
python-jose[cryptography]
passlib[bcrypt]
safe